WIP: race condition fix 2
I want to propose a different solution to the race
condition screening. What if we modified
MasterDetailData to include a path field that is
the path used to create the MasterDetailData? This
change would be in implementations of
IService.getData().  Then, ProjectEditor() could
dispense with getProjectIdFromPath() and
findNodeInTree() and just compare location with
data.path.

Your proposal to add a path field to
MasterDetailData is an excellent solution that
would simplify the code and make it more robust.
This approach would directly address the race
condition issue by providing a reliable way to
validate that the data matches the current path.